Positive basis algorithm
This is a set of programs designed to automatize the process of finding the "positive basis" for the nullspace of a stoichoimetric matrix. The positive basis is such that one can represent any element of the nullspace intersected with the positive orthant using a nonnegative linear combination of vectors. Often the positive basis has more elements than the dimensionality of the nullspace.